// src/GeneratedClient/models/WaitingListModelResultObject.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WaitingList
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { ValidationMessage } from './ValidationMessage';
import {
    ValidationMessageFromJSON,
    ValidationMessageFromJSONTyped,
    ValidationMessageToJSON,
    ValidationMessageToJSONTyped,
} from './ValidationMessage';
import type { WaitingListModel } from './WaitingListModel';
import {
    WaitingListModelFromJSON,
    WaitingListModelFromJSONTyped,
    WaitingListModelToJSON,
    WaitingListModelToJSONTyped,
} from './WaitingListModel';

/**
 * 
 * @export
 * @interface WaitingListModelResultObject
 */
export interface WaitingListModelResultObject {
    /**
     * 
     * @type {Array<WaitingListModel>}
     * @memberof WaitingListModelResultObject
     */
    records: Array<WaitingListModel>;
    /**
     * 
     * @type {Array<ValidationMessage>}
     * @memberof WaitingListModelResultObject
     */
    messages: Array<ValidationMessage>;
}

/**
 * Check if a given object implements the WaitingListModelResultObject interface.
 */
export function instanceOfWaitingListModelResultObject(value: object): value is WaitingListModelResultObject {
    if (!('records' in value) || value['records'] === undefined) return false;
    if (!('messages' in value) || value['messages'] === undefined) return false;
    return true;
}

export function WaitingListModelResultObjectFromJSON(json: any): WaitingListModelResultObject {
    return WaitingListModelResultObjectFromJSONTyped(json, false);
}

export function WaitingListModelResultObjectFromJSONTyped(json: any, ignoreDiscriminator: boolean): WaitingListModelResultObject {
    if (json == null) {
        return json;
    }
    return {
        
        'records': ((json['records'] as Array<any>).map(WaitingListModelFromJSON)),
        'messages': ((json['messages'] as Array<any>).map(ValidationMessageFromJSON)),
    };
}

export function WaitingListModelResultObjectToJSON(json: any): WaitingListModelResultObject {
    return WaitingListModelResultObjectToJSONTyped(json, false);
}

export function WaitingListModelResultObjectToJSONTyped(value?: WaitingListModelResultObject | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'records': ((value['records'] as Array<any>).map(WaitingListModelToJSON)),
        'messages': ((value['messages'] as Array<any>).map(ValidationMessageToJSON)),
    };
}